bugfix - do not resolve remote nick if it contains "/"
A user can set their upstream during a push by using the full URL
and adding --set-upstream. This adds a config value and bypasses all
the existing remote lookup checks in libgit2. This means an error
occurs when trying to resolve the URL to a remote nick.

We should return a null remote here. This might also be necessary to
fix in libgit2, but for now i'm just yoloing things.