Allow multiple concurrent filter streams
The current implementation of filters in libgit2sharp only allows for
one filter stream per type of filter at any given point in time. This
switches the storage of the necessary state from instance variables on
the Filter class to a dedicated state object which is tracked in a
dictionary keyed on the libgit2 stream pointer.